small fixes in 1_2, move qed.py and ro3.py content in drug_likeness.py
